+++ routers/userRouter.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const userController  = require('../controllers/userController');
+const { validToken } = require('../utils/auth')
 
 const router = express.Router();
 
@@ -7,5 +8,7 @@ router.post('/signup', userController.signUp);
 router.post('/check', userController.checkUser);
 
 router.post('/signin', userController.signIn);
+
+router.delete('/likes/deletelike', validToken, userController.deleteLike);
 
 module.exports = router;
