--- services/myPageService.js
@@ -1,17 +1,0 @@
-const myPageDao = require('../models/myPageDao');
-
-
-
-const deleteLike = async (userId, products) => {
-
-	return await myPageDao.deleteLike(userId, products);
-
-}
-
-
-
-
-
-module.exports = { 
-    deleteLike
-}
